+++ backend/src/server.js
@@ -64,6 +64,38 @@ app
 		],
 		credentials: true
 	}));
+
+
+const winston = require('winston')
+function getLogger(){
+	//Winston config
+	const myformat = winston.format.combine(
+		winston.format.colorize(),
+		winston.format.timestamp(),
+		winston.format.align(),
+		winston.format.printf(info => `${info.timestamp} ${info.level}: ${info.message}`)
+	);
+	const logConfiguration = {
+		transports: [
+			new winston.transports.Console({
+				level: 'debug',
+				format: myformat
+			}),
+			new winston.transports.File({
+				level: 'warn',
+				filename: './logs/backend_warn.log',
+				format: myformat
+			}),
+			new winston.transports.File({
+				level: 'debug',
+				filename: './logs/backend_debug.log',
+				format: myformat
+			})
+		]
+	};
+	return winston.createLogger(logConfiguration);
+}
+const logger = getLogger();
 app
 	.use(passport.initialize())
 	.use(passport.session())
@@ -71,6 +103,24 @@ app
 	.use(bodyParser.urlencoded({ limit: '100kb', extended: true }))
 	.use((_, __, next) => {
 		console.log('Time:', Date.now());
+		let current_datetime = new Date();
+		let formatted_date =
+			current_datetime.getFullYear() +
+			"-" +
+			(current_datetime.getMonth() + 1) +
+			"-" +
+			current_datetime.getDate() +
+			" " +
+			current_datetime.getHours() +
+			":" +
+			current_datetime.getMinutes() +
+			":" +
+			current_datetime.getSeconds();
+		let method = _.method;
+		let url = _.url;
+		let status = __.statusCode;
+		let log = `[${formatted_date}] ${method}:${url} ${status}`;
+		logger.debug(log)
 		next();
 	})
 	.use('/api/script', scriptRouter)
@@ -85,5 +135,6 @@ app
 	.get('/api', (_, res) => {
 		res.sendFile('htmlresponse/apistandartresponse.html', { root: __dirname });
 	});
+	
 
 module.exports = { app };
